Filter posts on category and tag pages.
The previous code was including every single post on every single tag and category page. I've switched it a bit so that we use the category's slug in the URL and pass both `slug` and `name` to the page context. Then we can use the `slug` to filter posts which have that category or tag.